requirements.txt
asgiref==3.3.4
click==8.0.1
cx-Oracle==8.2.1
fastapi==0.65.2
greenlet==1.1.0
h11==0.12.0
httptools==0.2.0
pydantic==1.8.2
PyJWT==2.1.0
python-dotenv==0.17.1
PyYAML==5.4.1
SQLAlchemy==1.4.18
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
